requirements.txt
fastapi==0.111.0
uvicorn==0.30.0
transformers==4.41.2
torch==2.2.2
pydantic==2.7.1
python-dotenv==1.0.1